refactor: move relevant code out of pinguim/qol.hpp
