+++ conf/meson.build
@@ -3,6 +3,14 @@ conf_data = configuration_data()
 # Keep in mind that the folder structure is replicated, thats why
 # conf.h.in must be in that separate folder.
 
+## Compiler stuff.
+cxxid = cxx.get_id()
+conf_data.set_quoted('cxx_id', cxx.get_id())
+conf_data.set('PINGUIM_CONF_COMPILER_IS_GCC',   cxxid == 'gcc')
+conf_data.set('PINGUIM_CONF_COMPILER_IS_CLANG', cxxid == 'clang')
+conf_data.set('PINGUIM_CONF_COMPILER_IS_MSVC',  cxxid == 'msvc')
+conf_data.set_quoted('cxx_version', cxx.version())
+
 # For OpenCV, we can check for stuff in <opencv2/cvconfig.h> at any time,
 # but we'd rather do this here than in the app, since doing so pollutes the
 # global namespace with a bunch of macros.
